+++ .env
@@ -2,21 +2,31 @@
 PORT=3000
 DATABASE_URL=postgres://knex@localhost/lynx
 
-# OIDC Config
-ISSUER_BASE_URL=http://localhost:8080/auth/realms/lynx-team/.well-known/openid-configuration
-CLIENT_ID=lynx-app
-BASE_URL=http://localhost:3000
-SECRET=replace me!
-
-## Point to user name/role fields in the JSON/JWT that the OIDC IdP returns on successful login
-USER_NAME_FIELD=preferred_username
-USER_ROLE=resource_access.lynx-app.roles[0]
-
 # Backend Config
 HOMEPAGE=https://nodejs.org
+BASE_URL=http://localhost:3000
+
+## OIDC Config
+### I know that it is possible to get `jwks_uri` from the OIDC "well known" endpoint;
+### however, it's such a pain in the ass to make an async call to that endpoint
+### before initializing literally anything else; so, for now, just hard-code it
+OIDC_JWKS_URI=http://localhost:8080/auth/realms/lynx-team/protocol/openid-connect/certs
+OIDC_ALGORITHMS=RS256
+OIDC_CLIENT_ID=lynx-app
+
+### Point to user name/role fields in the JWT that the OIDC IdP returns on successful login
+OIDC_USER_ROLE=resource_access.lynx-app.roles[0]
 
 ## How long to wait before the backend gives up scraping a link
 LINK_TIMEOUT=5 seconds
 
 # Frontend config
 REACT_APP_BASE_URL=$BASE_URL
+
+## OIDC Config
+REACT_APP_OIDC_CLIENT_ID=$OIDC_CLIENT_ID
+REACT_APP_OIDC_ISSUER=http://localhost:8080/auth/realms/lynx-team/.well-known/openid-configuration
+REACT_APP_REDIRECT_URI=http://localhost:3000/app
+
+REACT_APP_USER_NAME_FIELD=preferred_username
+REACT_APP_USER_ROLE=$OIDC_USER_ROLE
