+++ Makefile
@@ -6,7 +6,7 @@ up-all:
 	docker-compose up -d
 
 down:
-	docker-compose down
+	docker-compose down --remove-orphans
 
 logs:
 	docker-compose logs -f
