--- __mocks__/express-openid-connect.js
@@ -1,19 +1,0 @@
-exports.auth = () => (req, res, next) => {
-  const role = req.header('X-Mock-Role')
-
-  if (role) {
-    req.oidc = {
-      user: {
-        sub: 'testUser',
-        preferred_username: 'Test User',
-        resource_access: {
-          'lynx-app': {
-            roles: [role]
-          }
-        }
-      }
-    }
-  }
-
-  next()
-}
