+++ package.json
@@ -44,11 +44,13 @@
     "dotenv-flow": "^3.2.0",
     "express": "^4.17.1",
     "express-async-errors": "^3.1.1",
+    "express-jwt": "^6.0.0",
     "express-openid-connect": "^2.1.0",
     "express-query-boolean": "^2.0.0",
     "got": "^11.5.1",
     "helmet": "^4.0.0",
     "http-errors": "^1.7.3",
+    "jwks-rsa": "^1.12.2",
     "knex": "^0.21.15",
     "lodash": "^4.17.19",
     "metascraper": "^5.13.2",
@@ -91,10 +93,12 @@
     "husky": "^4.2.0",
     "jest-junit": "^12.0.0",
     "json-schema-faker": "^0.5.0-rcv.32",
+    "jwt-decode": "^3.1.2",
     "knex-cleaner": "^1.3.1",
     "lint-staged": "^10.0.1",
     "nodemon": "^2.0.2",
     "npm-run-all": "^4.1.5",
+    "oidc-client": "^1.11.3",
     "onchange": "^7.0.2",
     "pino-multi-stream": "^5.1.1",
     "pino-pretty": "^4.1.0",
