+++ routes/index.js
@@ -1,7 +1,7 @@
 const { Router } = require('express')
-const ensureLogin = require('../middlewares/ensure-login')
+const auth = require('../middlewares/authenticate')
 
 module.exports = Router()
-  .use('/api', ensureLogin, require('./api'))
+  .use('/api', auth, require('./api'))
   .use('/app', require('./app'))
   .use('/', require('./redirect'))
